# [Xamarin.Android.Build.Tasks] less string alloc in native assembly generation